karin: overlay: Enable config_suspendWhenScreenOffDueToProximity
Indicate whether to allow the device to suspend when the screen is off
due to the proximity sensor.  This resource should only be set to true
if the sensor HAL correctly handles the proximity sensor as a wake-up source.
Otherwise, the device may fail to wake out of suspend reliably.
The default is false.

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Ia21c7f69a0f26a49bce636628eaf9e66c109fa53